requirement.txt
numpy
opencv-python
matplotlib
pyserial
h5py
scipy
sklearn
bleak
shapely